backend/requirements.txt

Django==5.1.6
django-cors-headers==4.7.0
django-ninja==1.3.
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
twilio==9.4.6
